api/requirements.txt
fastapi-sqlalchemy
pydantic
psycopg2-binary
scikit-learn==0.24.2
aiofiles